fix: address review feedback for PWA refresh prompt
- Extract showUpdateBanner() to shared utility (DRY)
- Fix setInterval cleanup in ReloadPrompt (StrictMode safe)
- Fix never-resolving promise in lazyWithRetry (offline falls back to ErrorBoundary)
- Reduce SW update frequency from 1 hour to 4 hours
- Add changelog entry for v2.5.2

Addresses Codex and Claude bot review feedback.
